+++ src/pages/login/login.tsx
@@ -1,30 +1,13 @@
 import { Helmet } from 'react-helmet-async';
-import { Navigate, useLocation } from 'react-router-dom';
-import { useRef, FormEvent } from 'react';
+import { Link, Navigate, useLocation } from 'react-router-dom';
 
 import MemoizedHeader from '../../components/layout/header/header';
-import { loginAction } from '../../services/api-actions';
+import FormLogin from '../../components/blocks/form-login/form-login';
 import { AuthorizationStatus, AppRoute } from '../../components/consts';
-import { useAppDispatch,useAppSelector } from '../../store/useAppDispatch';
+import { useAppSelector } from '../../store/useAppDispatch';
 
 function Login(): JSX.Element {
   const location = useLocation();
-
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
-
-  const dispatch = useAppDispatch();
-
-  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      dispatch(loginAction({
-        login: loginRef.current.value,
-        password: passwordRef.current.value
-      }));
-    }
-  };
 
   const authorizationStatus = useAppSelector((state) => state.authorizationStatus) as AuthorizationStatus;
   if (authorizationStatus === AuthorizationStatus.Auth) {
@@ -43,42 +26,13 @@ function Login(): JSX.Element {
         <div className="page__login-container container">
           <section className="login">
             <h1 className="login__title">Sign in</h1>
-            <form
-              onSubmit={ handleSubmit }
-              className="login__form form"
-              action=""
-              method="post"
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  ref={ loginRef }
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  ref={ passwordRef }
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  required
-                />
-              </div>
-              <button className="login__submit form__submit button" type="submit">Sign in</button>
-            </form>
+            <FormLogin />
           </section>
           <section className="locations locations--login locations--current">
             <div className="locations__item">
-              <a className="locations__item-link" href="#">
+              <Link className="locations__item-link" to={ AppRoute.Main }>
                 <span>Amsterdam</span>
-              </a>
+              </Link>
             </div>
           </section>
         </div>
